added gate variable to components
